requirements.txt
python-dotenv
requests
flake8
pytest
